pkg/policy: use float64 for PolicyEntry type
This is (almost) user-facing API, so it's nice to use a float here so
that there is "always" arbitrary room for users to insert entries
between two existing entries.

This doesn't change the datapath, it just sorts the set of entries that
apply to a given endpoint and assigns an integer level to them.

+++ pkg/policy/types/policyentry.go
@@ -16,7 +16,7 @@ type PolicyEntry struct {
 	// take precedence over higher values. Rules having the default priority level 0 are
 	// considered first, then the rest of the rules, from the earliest to later priority levels.
 	// This is currently limited to 24 bits, i.e., max allowed priority is (1<<24-1).
-	Priority uint32
+	Priority float64
 
 	// Authentication specifies the cryptographic authentication required for the traffic to be
 	// allowed
